% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.func.plot.R
\name{beta.func.plot}
\alias{beta.func.plot}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
beta.func.plot(
  fitted,
  ngrid.x,
  ngrid.y,
  ngrid.t,
  boundary,
  boundary.t,
  beta.true.ind = TRUE,
  beta.true = NULL,
  coord.ratio = 1
)
}
\arguments{
\item{fitted}{Estimated STAR model.}

\item{ngrid.x}{Number of grid points in x axis.}

\item{ngrid.y}{Number of grid points in y axis.}

\item{ngrid.t}{Number of grid points in t axis.}

\item{boundary}{Boundary of spatial domain.}

\item{boundary.t}{Boundary of time domain.}

\item{beta.true.ind}{A logical number incidating whether to draw the true coefficient functions.}

\item{beta.true}{A list of true coefficient functions.}

\item{coord.ratio}{Aspect ratio, expressed as y / x.}
}
\value{
A list of plots
\item{beta.true.all}{A list of length \code{n.X} containing sequence of spatial plots
of each true coefficient functions.}
\item{beta.est.all}{A list of length \code{n.X} containing sequence of spatial plots
of each estiamted coefficient functions.}
}
\description{
This function generates sequence of spatial plots of estimated coefficient functions.
}
