% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.func.R, R/beta1.func.R
\name{beta1.func}
\alias{beta1.func}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
beta1.func(x, y, z)

beta1.func(x, y, z)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}

\item{t}{t coordinate}
}
\description{
This function is the coefficient function in simulation example.

This function is the coefficient function in simulation example.
}
