% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta3.func.R
\name{beta3.func}
\alias{beta3.func}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
beta3.func(x, y, z)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}
}
\description{
This function is the coefficient function in simulation example.
}
