% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.cv.R
\name{block.cv}
\alias{block.cv}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
block.cv(locations, nblock.x, nblock.y, nblock.t, nfold)
}
\arguments{
\item{locations}{A matrix with three columns stores the
locations of data points. The first two columns are spatial locations.}

\item{nblock.x}{Number of bins in x-dimension.}

\item{nblock.y}{Number of bins in y-dimension.}

\item{nblock.t}{Number of bins in t-dimension. The total number of
blocks is equal to \code{nblock.x * nblock.y * nblock.t}}

\item{nfold}{Number of folds in cross-validation.}
}
\value{
\item{Fold}{A list of length \code{nfold}. Each list contains the
index of data points in the testing set.}
}
\description{
This function generates testing sets for blockwise cross-validation
in spatiotemporal data analysis.
}
