% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.star.R
\name{eval.star}
\alias{eval.star}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
eval.star(
  fitted,
  beta.true = NULL,
  eta.true = NULL,
  alpha.true = NULL,
  data.grid = NULL
)
}
\arguments{
\item{fitted}{Estimated STAR model.}

\item{beta.true}{True beta functions.}

\item{eta.true}{True linear coefficients \eqn{\eta}.}

\item{alpha.true}{True \eqn{\alpha}.}

\item{data.grid}{Grid points to evaluate beta functions.}
}
\value{
\item{mse.alpha}{squared error of \eqn{\alpha}.}
\item{mse.eta}{squared error of \eqn{\eta}.}
\item{mise.beta}{integrated squared error of \eqn{\beta(s1, s2, t)}.}
\item{beta.hat.grid}{Estimated \eqn{\beta(s1, s2, t)} at grid points
\code{data.grid}.}
\item{beta.func.grid}{True \eqn{\beta(s1, s2, t)} at grid points
\code{data.grid}.}
}
\description{
This function evaluates the performance of estimated STAR model by
comparing with the true model.
}
