% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.plot.R
\name{raster.plot}
\alias{raster.plot}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
raster.plot(data, value.range, boundary, coord.ratio = 1)
}
\arguments{
\item{data}{The data to be displayed with columns:
\code{x}, \code{y}, and \code{value}.}

\item{value.range}{The range of value displayed on the plot.}

\item{boundary}{Boundary of the plot. The value out of boundary is \code{NA}.}

\item{coord.ratio}{Aspect ratio, expressed as y / x.}
}
\description{
This function generates raster plot using ggplot2.
}
