% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.star.eta.R
\name{se.star.eta}
\alias{se.star.eta}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
se.star.eta(
  Ystar,
  X,
  Z,
  W,
  energQ21,
  energQ22,
  Weight,
  Ai,
  PsiX0,
  mfit,
  Weight.type = "normal"
)
}
\description{
This function calculates standard deviation of parameters in STAR models.
}
