% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1.data.generating.R
\name{simu1.data.generating}
\alias{simu1.data.generating}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
simu1.data.generating(nS, nT, sigma)
}
\arguments{
\item{nS}{Number of points randomly generated from spatial domain.}

\item{nT}{Number of times that spatial points are observed across the time.}

\item{sigma}{Noise level.}
}
\value{
A list of five containing the data to fit model.
\item{Y}{response variable.}
\item{Z}{exogenous variables with constant linear
coefficients}
\item{X}{exogenous variables with varying linear
coefficients.}
\item{location}{A \code{nS * nT} by three matrix with locations of data points.}
}
\description{
This function generates simulation example for STVCM.
}
