% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu2.data.generating.R
\name{simu2.data.generating}
\alias{simu2.data.generating}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
simu2.data.generating(nS, nT, sigma, alpha, k = 10, boundary)
}
\arguments{
\item{nS}{Number of points randomly generated from spatial domain.}

\item{nT}{Number of times that spatial points are observed across the time.}

\item{sigma}{Noise level.}

\item{alpha}{\eqn{\alpha} in STAR model.}

\item{k}{The weight matrix is generated based on k-nearest neighbor.
The default value in this simulation example is 10.}

\item{boundary}{The boundary of the spatial domain.}
}
\value{
A list of five containing the data to fit model.
\item{Y}{response variable.}
\item{Z}{exogenous variables with constant linear
coefficients}
\item{X}{exogenous variables with varying linear
coefficients.}
\item{location}{A \code{nS * nT} by three matrix with locations of data points.}
\item{W}{weight matrix in STAR.}
}
\description{
This function generates simulation example for STAR with partially linear
varying coefficient model.
}
