% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.fit.R
\name{star.fit}
\alias{star.fit}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
star.fit(
  data,
  Lambda,
  V,
  Tri,
  d,
  r,
  time.knots,
  rho,
  time.bound,
  intercept = TRUE,
  det.func = Matrix::det,
  return.se = FALSE,
  Weight.type = "normal"
)
}
\arguments{
\item{data}{A list of five containing the data to fit model. \code{Y} is the
response variable. \code{Z} is exogenous variables with constant linear
coefficients. \code{X} is exogenous variables with varying linear
coefficients. \code{location} is the location of data points. \code{W} is the
weight matrix in STAR.}

\item{Lambda}{A matrix of tuning parameters of the smoothness penalty.}

\item{V}{The \code{N} by two matrix of verities of a triangulation,
where \code{N} is the number of vertices. Each row is the coordinates for a vertex.}

\item{Tri}{The triangulation matrix of dimension \code{nTr} by three,
where \code{nTr} is the number of triangles in the triangulation.
Each row is the indices of vertices in \code{V}.}

\item{d}{The degree of piecewise polynomials -- default is 2,
and usually \code{d} is greater than one. -1 represents piecewise constant.}

\item{r}{The smoothness parameter -- default is 1, and 0 \eqn{\le} \code{r} \eqn{<} \code{d}.}

\item{time.knots}{The vector of interior time.knots for univariate spline.}

\item{rho}{The order of univariate spline.}

\item{time.bound}{The vector of two. The boundary of univariate spline.}

\item{intercept}{A logical number indicating whether to include intercept term
in the model -- default is \code{TRUE}.}

\item{det.func}{A function used to calculate determinate of a matrix --
default is \code{det} in the R package \code{Matrix}.}

\item{return.se}{A logical number indicating whether to calculate standard deviation
of the linear estimators.}

\item{Weight.type}{The type of weight matrix. "\code{normal}" means
the weight matrix does not have special structure. "\code{upper}" means
the weight matrix is upper triangle matrix.}
}
\value{
\item{n.Z}{number of linear coefficients.}
\item{n.X}{number of varying coefficient functions.}
\item{best.lambda}{the selected smoothing tuning parameters.}
\item{theta.hat}{Estimated coefficents.}
\item{alpha.hat}{Estimated alpha in STAR.}
\item{gamma.hat}{Estimated tensor spline coefficients.}
\item{beta.hat}{Estimated tensor spline coefficients with matrix \eqn{Q_2}.}
\item{se.eta}{Standard deviation of \eqn{\alpha, \sigma^2} and linear coefficients.}
\item{Y.hat}{Estimated response variable.}
\item{mse}{Estimated \eqn{\sigma^2}.}
\item{mle}{MLE of fitted model.}
\item{MSE.Y}{Mean squared error of response variable.}
}
\description{
This function fits a spatiotemporal autoregressive (STAR) model
with exogenous variables.
}
