% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stvcm.fit.R
\name{stvcm.fit}
\alias{stvcm.fit}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
stvcm.fit(
  data,
  Lambda,
  V,
  Tri,
  d,
  r,
  time.knots,
  rho,
  time.bound,
  intercept.X = TRUE,
  intercept.Z = FALSE,
  return.se = FALSE
)
}
\arguments{
\item{data}{A list of five containing the data to fit model. \code{Y} is the
response variable. \code{Z} is exogenous variables with constant linear
coefficients. \code{X} is exogenous variables with varying linear
coefficients. \code{location} is the location of data points.}

\item{Lambda}{A matrix of tuning parameters of the smoothness penalty.}

\item{V}{The \code{N} by two matrix of verities of a triangulation,
where \code{N} is the number of vertices. Each row is the coordinates for a vertex.}

\item{Tri}{The triangulation matrix of dimension \code{nTr} by three,
where \code{nTr} is the number of triangles in the triangulation.
Each row is the indices of vertices in \code{V}.}

\item{d}{The degree of piecewise polynomials -- default is 2,
and usually \code{d} is greater than one. -1 represents piecewise constant.}

\item{r}{The smoothness parameter -- default is 1, and 0 \eqn{\le} \code{r} \eqn{<} \code{d}.}

\item{time.knots}{The vector of interior time.knots for univariate spline.}

\item{rho}{The order of univariate spline.
#' @param intercept.X A logical number indicating whether to include varying intercept term
in the model -- default is \code{TRUE}.}

\item{time.bound}{The vector of two. The boundary of univariate spline.}

\item{intercept.Z}{A logical number indicating whether to include constant intercept term
in the model -- default is \code{FALSE}.}

\item{return.se}{A logical number indicating whether to calculate standard deviation
of the linear estimators.}
}
\value{
\item{n.Z}{number of linear coefficients.}
\item{n.X}{number of varying coefficient functions.}
\item{best.lambda}{the selected smoothing tuning parameters.}
\item{theta.hat}{Estimated coefficents.}
\item{gamma.hat}{Estimated tensor spline coefficients.}
\item{beta.hat}{Estimated tensor spline coefficients with matrix \eqn{Q_2}.}
\item{se.eta}{Standard deviation of \eqn{\alpha, \sigma^2} and linear coefficients.}
\item{Y.hat}{Estimated response variable.}
\item{mse}{Estimated \eqn{\sigma^2}.}
\item{mle}{MLE of fitted model.}
\item{MSE.Y}{Mean squared error of response variable.}
}
\description{
This function fits a spatiotemporal varying coefficient model.
}
