% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.knn.R
\name{w.knn}
\alias{w.knn}
\title{STARX: spatiotemporal autoregressive model with exogenous variables}
\usage{
w.knn(coords, k = 10)
}
\description{
This function generates weight matrix based on k nearest neighbor.
}
